+++ Headers/IDEKit/IDEDistributionAutomaticSigningAssetsStepViewController.h
@@ -21,7 +21,7 @@
     long long _state;
 }
 
-+ (BOOL)skipStepForContext:(id)arg1 assistantDirection:(int)arg2;
++ (BOOL)skipStepForContext:(id)arg1 assistantDirection:(long long)arg2;
 + (id)keyPathsForValuesAffectingAdditionalActionTitle;
 + (id)keyPathsForValuesAffectingWantsPreviousAction;
 + (id)keyPathsForValuesAffectingWantsNextAction;
